+++ src/index.ts
@@ -7,4 +7,4 @@ export { ServiceCluster } from './service_cluster/index.js';
 
 export type { TaskDefinitions } from './types/definitions.js';
 export type { Nanolith } from './types/nanolith.js';
-export type { TaskWorkerOptions, ServiceWorkerOptions } from './types/workers.js';
+export type { TaskWorkerOptions as LaunchTaskOptions, ServiceWorkerOptions as LaunchServiceOptions } from './types/workers.js';