+++ src/service/service.ts
@@ -45,8 +45,6 @@ export class Service<Definitions extends TaskDefinitions> extends TypedEmitter<S
         // Only attach one listener onto the worker instead of attaching a listener for each task called.
         const taskHandler = (body: WorkerBaseMessageBody & { key: string }) => {
             this.#callbacks.forEach(({ resolve, reject }, key) => {
-                // Ignore all messages that aren't one of these two types.
-                if (body.type !== WorkerMessageType.CallError && body.type !== WorkerMessageType.CallReturn) return;
                 // If the message is for a call with a different key, also ignore the message.
                 if (body.key !== key) return;
 
@@ -68,7 +66,7 @@ export class Service<Definitions extends TaskDefinitions> extends TypedEmitter<S
         this.#worker.on('message', taskHandler);
 
         // Exit handler
-        const terminationHandler = () => {
+        this.#worker.once('exit', () => {
             this.#terminated = true;
             // Early cleanup of the callbacks map
             this.#callbacks.clear();
@@ -76,9 +74,7 @@ export class Service<Definitions extends TaskDefinitions> extends TypedEmitter<S
             this.emit('terminated');
             // Clean up task handler listener
             worker.off('message', taskHandler);
-        };
-
-        this.#worker.once('exit', terminationHandler);
+        });
     }
 
     /**